services/bff/pyproject.toml
[project]
name = "bff"
version = "0.1.0"
description = "Backend For Frontend (BFF) Service"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2",
    "django-ninja>=1.0",
    "django-cors-headers",
    "gunicorn",
    "httpx",
    "uvicorn",
    "psycopg[binary]",
    "dj-database-url",
    "redis",
    "django-redis",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app", "src/bff"]
